Move bq_cast in util module
